requirements.txt
fastapi
uvicorn
kubernetes
prometheus-api-client
numpy
pandas
pyyaml
python-dotenv